fixed CDash warning, was using wrong type for variable
git-svn-id: https://svn.opendtect.org/od/trunk@33776 a985855d-d384-4b45-ae33-31a410998c3c

+++ src/Strat/stratseqattrib.cc
@@ -439,7 +439,7 @@ int Strat::LayModAttribCalc::nextStep()
 	const int seqnb = mCast( int, seqidx_ );
 	if ( extrgates_.isEmpty() )
 	{
-	    const float ilay = seq.nearestLayerIdxAtZ( z );
+	    const int ilay = seq.nearestLayerIdxAtZ( z );
 	    const float halfwdth = seq.layers()[ilay]->thickness() / 2.f;
 	    zrg.setFrom( Interval<float>(z-halfwdth, z+halfwdth) );
 	}
